requirements.txt
# Automatically generated by https://github.com/damnever/pigar.
#--find-links https://download.pytorch.org/whl/torch_stable.html
torch
#--find-links https://download.pytorch.org/whl/torch_stable.html
torchvision
python-multipart
uvicorn[standard]==0.15.0
fastapi==0.87.0
Werkzeug
colorama
numpy
gunicorn
opencv-python-headless
pandas
requests
scipy
setuptools == 54.1.1
tqdm 
pillow 
click==8.0.3
itsdangerous==2.0.1
python-dotenv==0.19.1
six==1.16.0
typing-extensions==3.10.0.2
retrying==1.3.3
threadpoolctl==2.1.0
protobuf==3.18.0
pyyaml
matplotlib
seaborn
nltk
pytesseract
textblob
scikit_learn
flask-session
#pytz
requests
pymongo
flask_pymongo
dnspython
flask-bcrypt
pyjwt==2.0.0
datetime
passlib==1.7.4
elasticsearch
python-jose
python-multipart
fastapi-utils
python-docx
python-pptx
pdf2image
python-dotenv
pywebhdfs
imutils
easyocr
ultralytics